api/requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
autobahn==22.6.1
Automat==20.2.0
black==22.6.0
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
Markdown==3.4.1
msgpack==1.0.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.19.0
pytest-django==4.5.2
python-dotenv==0.20.0
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
zope.interface==5.4.0